oauth: remove oauth package (#35)
The oauth package was created as part of the semi-automated migration. During the migration the package was not used by server package. After additional testing the server package's implementation of the same functionality is now more complete.

This commit removes the oauth package to prevent confusion and simplify the codebase.

Fixes: #34

Signed-off-by: Benson Wong <benson@tailscale.com>